Improve import dependencies
