+++ dg/commands/cluster/install_cloud_pak_for_data.py
@@ -34,7 +34,7 @@ from dg.lib.cloud_pak_for_data.cpd_manager_factory import (
 
 @click.command(
     context_settings=dg.utils.click.create_default_map_from_dict(
-        dg.config.data_gate_configuration_manager.get_current_credentials()
+        dg.config.cluster_credentials_manager.cluster_credentials_manager.get_current_credentials()
     )
 )
 @click.option("--server", required=True, help="OpenShift server URL")
