+++ dg/commands/fyre/cluster/get_cluster_access_token.py
@@ -14,7 +14,7 @@
 
 import click
 
-import dg.config
+import dg.config.cluster_credentials_manager
 import dg.utils.click
 
 from dg.lib.fyre.cluster.fyre_cluster_factory import fyre_cluster_factory
@@ -22,7 +22,7 @@ from dg.lib.fyre.cluster.fyre_cluster_factory import fyre_cluster_factory
 
 @click.command(
     context_settings=dg.utils.click.create_default_map_from_dict(
-        dg.config.data_gate_configuration_manager.get_current_credentials()
+        dg.config.cluster_credentials_manager.cluster_credentials_manager.get_current_credentials()
     )
 )
 @click.option(
