+++ dg/commands/fyre/cluster/init_node_for_data_gate.py
@@ -17,13 +17,14 @@ import asyncio
 import click
 
 import dg.config
+import dg.config.cluster_credentials_manager
 import dg.utils.click
 import dg.utils.ssh
 
 
 @click.command(
     context_settings=dg.utils.click.create_default_map_from_dict(
-        dg.config.data_gate_configuration_manager.get_current_credentials()
+        dg.config.cluster_credentials_manager.cluster_credentials_manager.get_current_credentials()
     )
 )
 @click.option(
