+++ dg/commands/fyre/cluster/login.py
@@ -16,7 +16,7 @@ from typing import Union
 
 import click
 
-import dg.config
+import dg.config.cluster_credentials_manager
 import dg.utils.click
 
 from dg.lib.fyre.cluster.fyre_cluster_factory import fyre_cluster_factory
@@ -24,7 +24,7 @@ from dg.lib.fyre.cluster.fyre_cluster_factory import fyre_cluster_factory
 
 @click.command(
     context_settings=dg.utils.click.create_default_map_from_dict(
-        dg.config.data_gate_configuration_manager.get_current_credentials()
+        dg.config.cluster_credentials_manager.cluster_credentials_manager.get_current_credentials()
     )
 )
 @click.option("--cluster-name", required=True, help="cluster name")
